Update: 適応箇所が重複している際の処理を修正

+++ src/script/1.coffee
@@ -7,8 +7,12 @@ util = require "./util.js"
 output = require "./output.js"
 
 exports.output = (json, callback) ->
+  type = 1
   packVer = json["pack-version"]
   minecraftVer = json["minecraft-version"]
+
+  paths = []
+  escaped = [] # エスケープされたmodのオブジェクトの配列
 
   # 翻訳ファイルコピー
   for mod in json.mods
@@ -30,15 +34,27 @@ exports.output = (json, callback) ->
         # pathが配列の時
         if util.isArray(mod.path)
           for key, val of mod.path
-            fs.copySync("../../../MinecraftModLangFiles/#{mod.name}/#{mod.version}/#{key}",
-                        "../../temp/#{minecraftVer} - #{packVer} - 1/#{val}")
+            if not util.existInArray(paths, val)
+              fs.copySync("../../../MinecraftModLangFiles/#{mod.name}/#{mod.version}/#{key}",
+                          "../../temp/#{minecraftVer} - #{packVer} - 1/#{val}")
+            else
+              fs.copySync("../../../MinecraftModLangFiles/#{mod.name}/#{mod.version}/#{key}",
+                          "../../temp/#{minecraftVer} - #{packVer} - 1/#{val} - #{mod.version}")
+              escaped.push(mod)
+            paths.push(val)
         else if util.isString(mod.path)
-          fs.copySync("../../../MinecraftModLangFiles/#{mod.name}/#{mod.version}",
-                      "../../temp/#{minecraftVer} - #{packVer} - 1/#{mod.path}")
+          if not util.existInArray(paths, mod.path)
+            fs.copySync("../../../MinecraftModLangFiles/#{mod.name}/#{mod.version}",
+                        "../../temp/#{minecraftVer} - #{packVer} - 1/#{mod.path}")
+          else
+            fs.copySync("../../../MinecraftModLangFiles/#{mod.name}/#{mod.version}",
+                        "../../temp/#{minecraftVer} - #{packVer} - 1/#{mod.path} - #{mod.version}")
+            escaped.push(mod)
+          paths.push(mod.path)
 
   # readme作成
   title = "/lang_S_#{minecraftVer}(リソースパック型)///////////////////////by S/////"
-  readme = output.makeReadmeText(json, 1, title)
+  readme = output.makeReadmeText(json, type, title, escaped)
   fs.writeFileSync("../../temp/#{minecraftVer} - #{packVer} - 1/Readme - 導入前に読んでください.txt",
                    readme)
 
@@ -49,5 +65,5 @@ exports.output = (json, callback) ->
               "../../temp/#{minecraftVer} - #{packVer} - 1/pack.png")
 
   # zip
-  output.zipUp("S_lang_files_#{minecraftVer}_#{packVer}.zip", json, 1, callback)
+  output.zipUp("S_lang_files_#{minecraftVer}_#{packVer}.zip", json, type, callback)
   return
